requirements.txt
fastapi==0.104.1
httpx==0.24.0
neo4j==5.15.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-cov==4.1.0
uvicorn==0.20.0
sqlalchemy==2.0.23
passlib==1.7.4
boto3==1.33.8
psycopg2-binary==2.9.9
bcrypt==4.0.1
openai==1.3.7
chevron==0.14.0
alembic==1.13.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
regopy==0.3.10
requests==2.31.0
python-statemachine[diagrams]==2.1.2
pydot==2.0.0
zep-python==1.5.0
pyotp==2.9.0
